Выделяет компонент рейтинга
